// src/app/models/pastor.model.ts
export class PastorModel {
  id?: number;
  name: string;
  category: string;
  year: string;
  area: string;
  membresia?: string;
  lugardeministerio?: string;
  titulos?: string;
  // constructor( name:string, lastname:string, cargo:string, username:string, password:string, type:string ){

  // }
}